Move Client Dashboard pages out of src/components
