+++ src/modules/assessments/assessmentTypes.ts
@@ -1,0 +1,5 @@
+import { GetClientAssessmentsQuery } from '@/types/gqlTypes';
+
+export type ClientAssessmentType = NonNullable<
+  NonNullable<GetClientAssessmentsQuery['client']>['assessments']
+>['nodes'][0];
